requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.13
pycryptodome==3.16.0
requests==2.31.0
urllib3==2.0.3
yarg==0.1.9
